rearchitect netproxy as shim
